psc: fix bug on 32-bit systems
long long is the data type to use if you want
a 64-bit integer on both 32-bit and 64-bit systems

+++ src/include/psc_particles_c.h
@@ -16,7 +16,7 @@ typedef struct psc_particle_c {
   particle_c_real_t qni;
   particle_c_real_t mni;
   particle_c_real_t wni;
-  long              kind; // 64 bits to match the other members, for bnd exchange
+  long long         kind; // 64 bits to match the other members, for bnd exchange
 } particle_c_t;
 
 struct psc_particles_c {
